+++ CHANGELOG.md
@@ -9,3 +9,8 @@ All notable changes to the "VSSticky" extension will be documented in this file.
 ## v1.0.1
 
 - Sorted Notes in the `All Notes View` according to the path to logically group the notes.
+
+## v1.0.2
+
+- Added a TODO list command to consolidate all the notes as a single Markdown preview.
+- Added commands to Import and Export the sticky notes.